frontend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.0
jinja2==3.1.6
pydantic==2.5.0

# Pinecone - versão que funciona localmente
pinecone>=7.0.0

# Google AI
google-generativeai==0.8.5
